[SVExtractTestCode] Extend extraction through some non-dataflow ops.
This extends the set of what is extracted to be able to see through
wires and instances, using the same backward slice algorithm. When it
can be shown that the fan-out of an instance is only used in the
extracted slice, the instance and its fan-out are removed, rather than
left behind. Other ops are presumably removed later by
canonicalization.

This allows some cases of instances that only existed to feed into the
test code to now be removed, which prevents leaving behind dead code
that may trip lint warnings.